Refactor: 슬라이더, 카테고리 필터링, 게시물 목록 등을 컴포넌트화 하고, CategoryContext 도입 - FeaturedSlider.tsx : 최근 3개 게시물 슬라이더 - CategoryFilter.tsx: 카테고리 필터링 - PostGrid.tsx: 게시물 목록 그리드
